test(api): regroup blocked user and temporarily blocked user seeds for efficiency in the same file
since it's for testing the same code.
